workshop assignment week 3

+++ routes/campsiteRouter.js
@@ -15,7 +15,7 @@ campsiteRouter.route('/')
             })
             .catch(err => next(err));
     })
-    .post(authenticate.verifyUser, (req, res, next) => {
+    .post(authenticate.verifyUser, authenticate.verifyAdmin, (req, res, next) => {
         console.log(res.body);
         Campsite.create(req.body)
             .then(campsite => {
@@ -30,7 +30,7 @@ campsiteRouter.route('/')
         res.statusCode = 403;
         res.end('PUT operation not supported on /campsites');
     })
-    .delete(authenticate.verifyUser, (req, res, next) => {
+    .delete(authenticate.verifyUser, authenticate.verifyAdmin, (req, res, next) => {
         Campsite.deleteMany()
             .then(response => {
                 res.statusCode = 200;
@@ -40,22 +40,22 @@ campsiteRouter.route('/')
             .catch(err => next(err));
     });
 
-    campsiteRouter.route('/:campsiteId')
+campsiteRouter.route('/:campsiteId')
     .get((req, res, next) => {
         Campsite.findById(req.params.campsiteId)
-        .populate('comments.author')
-        .then(campsite => {
-            res.statusCode = 200;
-            res.setHeader('Content-Type', 'application/json');
-            res.json(campsite);
-        })
-        .catch(err => next(err));
+            .populate('comments.author')
+            .then(campsite => {
+                res.statusCode = 200;
+                res.setHeader('Content-Type', 'application/json');
+                res.json(campsite);
+            })
+            .catch(err => next(err));
     })
     .post(authenticate.verifyUser, (req, res) => {
         res.statusCode = 403;
         res.end(`POST operation not supported on /campsites/${req.params.campsiteId}`);
     })
-    .put(authenticate.verifyUser, (req, res, next) => {
+    .put(authenticate.verifyUser, authenticate.verifyAdmin, (req, res, next) => {
         Campsite.findByIdAndUpdate(req.params.campsiteId, {
             $set: req.body
         }, { new: true })
@@ -66,7 +66,7 @@ campsiteRouter.route('/')
             })
             .catch(err => next(err));
     })
-    .delete(authenticate.verifyUser, (req, res, next) => {
+    .delete(authenticate.verifyUser, authenticate.verifyAdmin, (req, res, next) => {
         Campsite.findByIdAndDelete(req.params.campsiteId)
             .then(response => {
                 res.statusCode = 200;
@@ -76,28 +76,28 @@ campsiteRouter.route('/')
             .catch(err => next(err));
     });
 
-    campsiteRouter.route('/:campsiteId/comments')
+campsiteRouter.route('/:campsiteId/comments')
     .get((req, res, next) => {
         Campsite.findById(req.params.campsiteId)
-        .populate('comments.author')
-        .then(campsite => {
-            if (campsite) {
-                res.statusCode = 200;
-                res.setHeader('Content-Type', 'application/json');
-                res.json(campsite.comments);
-            } else {
-                err = new Error(`Campsite ${req.params.campsiteId} not found`);
-                err.status = 404;
-                return next(err);
-            }
-        })
-        .catch(err => next(err));
+            .populate('comments.author')
+            .then(campsite => {
+                if (campsite) {
+                    res.statusCode = 200;
+                    res.setHeader('Content-Type', 'application/json');
+                    res.json(campsite.comments);
+                } else {
+                    err = new Error(`Campsite ${req.params.campsiteId} not found`);
+                    err.status = 404;
+                    return next(err);
+                }
+            })
+            .catch(err => next(err));
     })
     .post(authenticate.verifyUser, (req, res, next) => {
         Campsite.findById(req.params.campsiteId)
-        .then(campsite => {
-            if (campsite) {
-                req.body.author = req.user._id;
+            .then(campsite => {
+                if (campsite) {
+                    req.body.author = req.user._id;
                     campsite.comments.push(req.body);
                     campsite.save()
                         .then(campsite => {
@@ -118,7 +118,7 @@ campsiteRouter.route('/')
         res.statusCode = 403;
         res.end(`PUT operation not supported on /campsites/${req.params.campsiteId}/comments`);
     })
-    .delete(authenticate.verifyUser, (req, res, next) => {
+    .delete(authenticate.verifyUser, authenticate.verifyAdmin, (req, res, next) => {
         Campsite.findById(req.params.campsiteId)
             .then(campsite => {
                 if (campsite) {
@@ -144,7 +144,7 @@ campsiteRouter.route('/')
 campsiteRouter.route('/:campsiteId/comments/:commentId')
     .get((req, res, next) => {
         Campsite.findById(req.params.campsiteId)
-        .populate('comments.author')
+            .populate('comments.author')
             .then(campsite => {
                 if (campsite && campsite.comments.id(req.params.commentId)) {
                     res.statusCode = 200;
@@ -170,19 +170,26 @@ campsiteRouter.route('/:campsiteId/comments/:commentId')
         Campsite.findById(req.params.campsiteId)
             .then(campsite => {
                 if (campsite && campsite.comments.id(req.params.commentId)) {
-                    if (req.body.rating) {
-                        campsite.comments.id(req.params.commentId).rating = req.body.rating;
+                    // Check if user is the author of the comment
+                    if (campsite.comments.id(req.params.commentId).author.equals(req.user._id)) {
+                        if (req.body.rating) {
+                            campsite.comments.id(req.params.commentId).rating = req.body.rating;
+                        }
+                        if (req.body.text) {
+                            campsite.comments.id(req.params.commentId).text = req.body.text;
+                        }
+                        campsite.save()
+                            .then(campsite => {
+                                res.statusCode = 200;
+                                res.setHeader('Content-Type', 'application/json');
+                                res.json(campsite);
+                            })
+                            .catch(err => next(err));
+                    } else {
+                        err = new Error('You are not authorized to update this comment!');
+                        err.status = 403;
+                        return next(err);
                     }
-                    if (req.body.text) {
-                        campsite.comments.id(req.params.commentId).text = req.body.text;
-                    }
-                    campsite.save()
-                        .then(campsite => {
-                            res.statusCode = 200;
-                            res.setHeader('Content-Type', 'application/json');
-                            res.json(campsite);
-                        })
-                        .catch(err => next(err));
                 } else if (!campsite) {
                     err = new Error(`Campsite ${req.params.campsiteId} not found`);
                     err.status = 404;
@@ -199,14 +206,21 @@ campsiteRouter.route('/:campsiteId/comments/:commentId')
         Campsite.findById(req.params.campsiteId)
             .then(campsite => {
                 if (campsite && campsite.comments.id(req.params.commentId)) {
-                    campsite.comments.id(req.params.commentId).deleteOne();
-                    campsite.save()
-                        .then(campsite => {
-                            res.statusCode = 200;
-                            res.setHeader('Content-Type', 'application/json');
-                            res.json(campsite);
-                        })
-                        .catch(err => next(err));
+                    // Check if user is the author of the comment
+                    if (campsite.comments.id(req.params.commentId).author.equals(req.user._id)) {
+                        campsite.comments.id(req.params.commentId).deleteOne();
+                        campsite.save()
+                            .then(campsite => {
+                                res.statusCode = 200;
+                                res.setHeader('Content-Type', 'application/json');
+                                res.json(campsite);
+                            })
+                            .catch(err => next(err));
+                    } else {
+                        err = new Error('You are not authorized to delete this comment!');
+                        err.status = 403;
+                        return next(err);
+                    }
                 } else if (!campsite) {
                     err = new Error(`Campsite ${req.params.campsiteId} not found`);
                     err.status = 404;
